# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/autoCode/createTemp": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "自动代码模板",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    },
                    {
                        "description": "创建自动代码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.AutoCodeStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"创建成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/autoCode/getColumn": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "获取当前表所有字段",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取当前表所有字段",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/autoCode/getDatabase": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "获取当前所有数据库",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取当前所有数据库",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/autoCode/getTables": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "获取当前数据库所有表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取当前数据库所有表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/autoCode/preview": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "预览创建后的代码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    },
                    {
                        "description": "预览创建代码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.AutoCodeStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "预览创建后的代码",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "system.AutoCodeStruct": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "description": "Struct简称",
                    "type": "string"
                },
                "autoCreateApiToSql": {
                    "description": "是否自动创建api",
                    "type": "boolean"
                },
                "autoMoveFile": {
                    "description": "是否自动移动文件",
                    "type": "boolean"
                },
                "description": {
                    "description": "Struct中文名称",
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.Field"
                    }
                },
                "humpPackageName": {
                    "description": "go文件名称",
                    "type": "string"
                },
                "packageName": {
                    "description": "文件名称",
                    "type": "string"
                },
                "structName": {
                    "description": "Struct名称",
                    "type": "string"
                },
                "tableName": {
                    "description": "表名",
                    "type": "string"
                }
            }
        },
        "system.Field": {
            "type": "object",
            "properties": {
                "columnName": {
                    "description": "数据库字段",
                    "type": "string"
                },
                "comment": {
                    "description": "数据库字段描述",
                    "type": "string"
                },
                "dataTypeLong": {
                    "description": "数据库字段长度",
                    "type": "string"
                },
                "dictType": {
                    "description": "字典",
                    "type": "string"
                },
                "fieldDesc": {
                    "description": "中文名",
                    "type": "string"
                },
                "fieldJson": {
                    "description": "FieldJson",
                    "type": "string"
                },
                "fieldName": {
                    "description": "Field名",
                    "type": "string"
                },
                "fieldSearchType": {
                    "description": "搜索条件",
                    "type": "string"
                },
                "fieldType": {
                    "description": "Field数据类型",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}